+++ CHANGES.md
@@ -35,6 +35,7 @@ Changes in 12.5
 - JS remove inline code from `_makePaymentsCommentsInput()` in Accounting/functions.inc.php & MakePaymentsCommentsInput.js
 - JS remove inline code from program in Administration.php & Administration.js
 - JS remove inline code from program in FixDailyAttendance.php & FixDailyAttendance.js
+- JS remove inline code from RegistrationAdminContactEnable() in RegistrationAdmin.fnc.php & RegistrationAdminContactEnable.js
 
 Changes in 12.4.3
 -----------------